+++ tests/unit/spinner/spinner.spec.js
@@ -8,6 +8,17 @@ describe('Spinner', () => {
     const shallow = enzyme.shallow;
     const shallowToJson = enzymeToJson.shallowToJson;
 
+    it('should not modify mods', () => {
+      const mods = ['test'];
+      const wrapper = shallow(
+        <Spinner mods={mods} size="xs">Extra small</Spinner>
+      );
+
+      expect(mods.length).toEqual(1);
+      expect(mods[0]).toEqual('test');
+      expect(shallowToJson(wrapper)).toMatchSnapshot();
+    });
+
     it('should return base class with mods for strings as class and string as mods', () => {
       const wrapper = shallow(
         <Spinner size="xs">Extra small</Spinner>